feature - add temp resource file.
